Refactor ctxrw to wrappers
